q-e/renamed enums ErrorMessage.ts -> ErrorMessage.enum.ts and Patterns
