+++ src/app/utils/Patterns.enum..ts
